+++ test/avoid_barrel_cycle_test.dart
@@ -1,5 +1,4 @@
 /// Tests for the avoid_barrel_cycle rule
-library;
 
 import 'package:analyzer_testing/analysis_rule/analysis_rule.dart';
 import 'package:barrel_file_lints/barrel_file_lints.dart';